✨ feat(synex)!: remove totokens bound
We introduced last commit a generic param. This param specifies the
ultimate desired type, after all is unwrapped.  This allowed to build
arbitrary intermediary Iter types through iter composition while not
setting in stone the ultimate return type.

Still the previous implementation did not distinguish between early
return because of an error and normal ebd of iteration return. This is
fixed now, but it's a breaking change.

BREAKING CHANGE: Wrap yielded results with Result<...>