✨ feat(synex): testsuite
Added a simple Extract impl for DeriveInput and a straightforward
"quoted" test which synex can't run.

What actually matters is the creation and use of an external package
to run the test. The whole complicated setup is made necessary by the
fact that synex is not allowed to implement any proc macros and
therefore can't run any test.

But it *CAN* define its tests within quotes and pass them to an
auxiliary package whose sole purpose is to run them.