✨ feat(task): parse backlog.data format
Lines are JSON objects.

All keys are the same as in pending.data, except for annotations,
which is an array of {"entry":"dt","description":"..."}.

Datetimes are not timestamps but "yyyyMMddThhmmssZ".

imask is a number, not a string.

All key cases are moved from parse mod to task mod, to allow
reusability for JSON object key/value.